Add Client Sign Helper
